+++ pkg/certificates/authorizations/authorizations_test.go
@@ -22,7 +22,7 @@ import (
 
 	"storj.io/storj/internal/testcontext"
 	"storj.io/storj/internal/testidentity"
-	"storj.io/storj/pkg/certificates/certificates_client"
+	"storj.io/storj/pkg/certificates/certificatesclient"
 	"storj.io/storj/pkg/identity"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/peertls/tlsopts"
@@ -597,7 +597,7 @@ func TestNewClient(t *testing.T) {
 	clientTransport := transport.NewClient(tlsOptions)
 
 	t.Run("Basic", func(t *testing.T) {
-		client, err := certificates_client.NewClient(ctx, clientTransport, listener.Addr().String())
+		client, err := certificatesclient.NewClient(ctx, clientTransport, listener.Addr().String())
 		assert.NoError(t, err)
 		assert.NotNil(t, client)
 
@@ -614,7 +614,7 @@ func TestNewClient(t *testing.T) {
 		pbClient := pb.NewCertificatesClient(conn)
 		require.NotNil(t, pbClient)
 
-		client, err := certificates_client.NewClientFrom(pbClient)
+		client, err := certificatesclient.NewClientFrom(pbClient)
 		assert.NoError(t, err)
 		assert.NotNil(t, client)
